refactor: abstract away the remote
