+++ git-cliff-core/src/remote/gitlab.rs
@@ -1,37 +1,11 @@
 use crate::config::Remote;
-use crate::error::{
-	Error,
-	Result,
-};
-use futures::{
-	future,
-	stream,
-	StreamExt,
-};
-use http_cache_reqwest::{
-	CACacheManager,
-	Cache,
-	CacheMode,
-	HttpCache,
-	HttpCacheOptions,
-};
-use reqwest::header::{
-	HeaderMap,
-	HeaderValue,
-};
-use reqwest::Client;
-use reqwest_middleware::{
-	ClientBuilder,
-	ClientWithMiddleware,
-};
-use secrecy::ExposeSecret;
-use serde::de::DeserializeOwned;
+use crate::error::*;
+use reqwest_middleware::ClientWithMiddleware;
 use serde::{
 	Deserialize,
 	Serialize,
 };
 use std::env;
-use std::time::Duration;
 
 use super::*;
 
@@ -49,13 +23,6 @@ pub const FINISHED_FETCHING_MSG: &str = "Done fetching GitLab data.";
 
 /// Template variables related to this remote.
 pub const TEMPLATE_VARIABLES: &[&str] = &["gitlab", "commit.gitlab"];
-
-/// Returns the GitLab API url either from environment or from default value.
-fn get_api_url() -> String {
-	env::var(GITLAB_API_URL_ENV)
-		.ok()
-		.unwrap_or_else(|| GITLAB_API_URL.to_string())
-}
 
 /// https://docs.gitlab.com/ee/api/projects.html#get-single-project
 /// Representation of a single GitLab Project.
@@ -78,8 +45,8 @@ pub struct GitLabProject {
 }
 
 impl RemoteEntry for GitLabProject {
-	fn url(_id: i64, repo: &str, owner: &str, _page: i32) -> String {
-		format!("{}/projects/{}%2F{}", get_api_url(), owner, repo)
+	fn url(_id: i64, api_url: &str, remote: &Remote, _page: i32) -> String {
+		format!("{}/projects/{}%2F{}", api_url, remote.owner, remote.repo)
 	}
 	fn buffer_size() -> usize {
 		1
@@ -129,13 +96,12 @@ impl RemoteCommit for GitLabCommit {
 }
 
 impl RemoteEntry for GitLabCommit {
-	fn url(id: i64, _repo: &str, _owner: &str, page: i32) -> String {
+	fn url(id: i64, api_url: &str, _remote: &Remote, page: i32) -> String {
 		let commit_page = page + 1;
 		format!(
 			"{}/projects/{}/repository/commits?per_page={MAX_PAGE_SIZE}&\
 			 page={commit_page}",
-			get_api_url(),
-			id
+			api_url, id
 		)
 	}
 	fn buffer_size() -> usize {
@@ -194,12 +160,11 @@ impl RemotePullRequest for GitLabMergeRequest {
 }
 
 impl RemoteEntry for GitLabMergeRequest {
-	fn url(id: i64, _repo: &str, _owner: &str, page: i32) -> String {
+	fn url(id: i64, api_url: &str, _remote: &Remote, page: i32) -> String {
 		format!(
 			"{}/projects/{}/merge_requests?per_page={MAX_PAGE_SIZE}&page={page}&\
 			 state=merged",
-			get_api_url(),
-			id
+			api_url, id
 		)
 	}
 
@@ -239,131 +204,40 @@ pub struct GitLabReference {
 /// HTTP client for handling GitLab REST API requests.
 #[derive(Debug, Clone)]
 pub struct GitLabClient {
-	/// Owner of the repository.
-	owner:  String,
-	/// GitLab repository.
-	repo:   String,
+	/// Remote.
+	remote: Remote,
 	/// HTTP client.
 	client: ClientWithMiddleware,
 }
 
-/// Constructs a GitLab client from the remote configuration.
+/// Constructs a GitHub client from the remote configuration.
 impl TryFrom<Remote> for GitLabClient {
 	type Error = Error;
 	fn try_from(remote: Remote) -> Result<Self> {
-		if !remote.is_set() {
-			return Err(Error::RemoteNotSetError);
-		}
-		let mut headers = HeaderMap::new();
-		headers.insert(
-			reqwest::header::ACCEPT,
-			HeaderValue::from_static("application/json"),
-		);
-		if let Some(token) = remote.token {
-			headers.insert(
-				reqwest::header::AUTHORIZATION,
-				format!("Bearer {}", token.expose_secret()).parse()?,
-			);
-		}
-		headers.insert(reqwest::header::USER_AGENT, USER_AGENT.parse()?);
-		let client = Client::builder()
-			.timeout(Duration::from_secs(REQUEST_TIMEOUT))
-			.tcp_keepalive(Duration::from_secs(REQUEST_KEEP_ALIVE))
-			.default_headers(headers)
-			.build()?;
-		let client = ClientBuilder::new(client)
-			.with(Cache(HttpCache {
-				mode:    CacheMode::Default,
-				manager: CACacheManager {
-					path: dirs::cache_dir()
-						.ok_or_else(|| {
-							Error::DirsError(String::from(
-								"failed to find the user's cache directory",
-							))
-						})?
-						.join(env!("CARGO_PKG_NAME")),
-				},
-				options: HttpCacheOptions::default(),
-			}))
-			.build();
 		Ok(Self {
-			owner: remote.owner,
-			repo: remote.repo,
-			client,
+			client: create_remote_client(&remote, "application/json")?,
+			remote,
 		})
 	}
 }
 
+impl RemoteClient for GitLabClient {
+	fn api_url() -> String {
+		env::var(GITLAB_API_URL_ENV)
+			.ok()
+			.unwrap_or_else(|| GITLAB_API_URL.to_string())
+	}
+
+	fn remote(&self) -> Remote {
+		self.remote.clone()
+	}
+
+	fn client(&self) -> ClientWithMiddleware {
+		self.client.clone()
+	}
+}
+
 impl GitLabClient {
-	/// Retrieves a single page of entries.
-	async fn get_entries_with_page<T: DeserializeOwned + RemoteEntry>(
-		&self,
-		project_id: i64,
-		page: i32,
-	) -> Result<Vec<T>> {
-		let url = T::url(project_id, &self.repo, &self.owner, page);
-		debug!("Sending request to: {url}");
-		let response = self.client.get(&url).send().await?;
-		let response_text = if response.status().is_success() {
-			let text = response.text().await?;
-			trace!("Response: {:?}", text);
-			text
-		} else {
-			let text = response.text().await?;
-			error!("Request error: {}", text);
-			text
-		};
-		let response = serde_json::from_str::<Vec<T>>(&response_text)?;
-		if response.is_empty() {
-			Err(Error::PaginationError(String::from("end of entries")))
-		} else {
-			Ok(response)
-		}
-	}
-
-	/// Fetches the GitLab API returns the given entry.
-	async fn fetch<T: DeserializeOwned + RemoteEntry>(
-		&self,
-		project_id: i64,
-	) -> Result<Vec<T>> {
-		let entries: Vec<Vec<T>> = stream::iter(0..)
-			.map(|i| self.get_entries_with_page(project_id, i))
-			.buffered(T::buffer_size())
-			.take_while(|page| {
-				if let Err(e) = page {
-					debug!("Error while fetching page: {:?}", e);
-				}
-				future::ready(page.is_ok())
-			})
-			.map(|page| match page {
-				Ok(v) => v,
-				Err(ref e) => {
-					log::error!("{:#?}", e);
-					page.expect("failed to fetch page: {}")
-				}
-			})
-			.collect()
-			.await;
-		Ok(entries.into_iter().flatten().collect())
-	}
-
-	/// Retrieves a single object.
-	async fn get_entry<T: DeserializeOwned + RemoteEntry>(&self) -> Result<T> {
-		let url = T::url(0, &self.repo, &self.owner, 1);
-		debug!("Sending request to: {url}");
-		let response = self.client.get(&url).send().await?;
-		let response_text = if response.status().is_success() {
-			let text = response.text().await?;
-			trace!("Response: {:?}", text);
-			text
-		} else {
-			let text = response.text().await?;
-			error!("Request error: {}", text);
-			text
-		};
-		Ok(serde_json::from_str::<T>(&response_text)?)
-	}
-
 	/// Fetches the GitLab API and returns the pull requests.
 	pub async fn get_project(&self) -> Result<GitLabProject> {
 		self.get_entry::<GitLabProject>().await
