+++ git-cliff-core/src/remote/mod.rs
@@ -6,7 +6,35 @@ pub mod github;
 #[cfg(feature = "gitlab")]
 pub mod gitlab;
 
+use crate::config::Remote;
+use crate::error::{
+	Error,
+	Result,
+};
 use dyn_clone::DynClone;
+use futures::{
+	future,
+	stream,
+	StreamExt,
+};
+use http_cache_reqwest::{
+	CACacheManager,
+	Cache,
+	CacheMode,
+	HttpCache,
+	HttpCacheOptions,
+};
+use reqwest::header::{
+	HeaderMap,
+	HeaderValue,
+};
+use reqwest::Client;
+use reqwest_middleware::{
+	ClientBuilder,
+	ClientWithMiddleware,
+};
+use secrecy::ExposeSecret;
+use serde::de::DeserializeOwned;
 use serde::{
 	Deserialize,
 	Serialize,
@@ -15,6 +43,7 @@ use std::hash::{
 	Hash,
 	Hasher,
 };
+use std::time::Duration;
 
 /// User agent for interacting with the GitHub API.
 ///
@@ -31,38 +60,10 @@ pub(crate) const REQUEST_KEEP_ALIVE: u64 = 60;
 /// Maximum number of entries to fetch in a single page.
 pub(crate) const MAX_PAGE_SIZE: usize = 100;
 
-/// Metadata of a remote release.
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
-pub struct RemoteReleaseMetadata {
-	/// Contributors.
-	pub contributors: Vec<RemoteContributor>,
-}
-
-/// Representation of a remote contributor.
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
-pub struct RemoteContributor {
-	/// Username.
-	pub username:      Option<String>,
-	/// Title of the pull request.
-	pub pr_title:      Option<String>,
-	/// The pull request that the user created.
-	pub pr_number:     Option<i64>,
-	/// Labels of the pull request.
-	pub pr_labels:     Vec<String>,
-	/// Whether if the user contributed for the first time.
-	pub is_first_time: bool,
-}
-
-impl Hash for RemoteContributor {
-	fn hash<H: Hasher>(&self, state: &mut H) {
-		self.username.hash(state);
-	}
-}
-
 /// Trait for handling the different entries returned from the remote.
 pub(crate) trait RemoteEntry {
 	/// Returns the API URL for fetching the entries at the specified page.
-	fn url(project_id: i64, repo: &str, owner: &str, page: i32) -> String;
+	fn url(project_id: i64, api_url: &str, remote: &Remote, page: i32) -> String;
 	/// Returns the request buffer size.
 	fn buffer_size() -> usize;
 }
@@ -94,6 +95,157 @@ dyn_clone::clone_trait_object!(RemotePullRequest);
 /// Result of a remote metadata.
 pub type RemoteMetadata =
 	(Vec<Box<dyn RemoteCommit>>, Vec<Box<dyn RemotePullRequest>>);
+
+/// Metadata of a remote release.
+#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
+pub struct RemoteReleaseMetadata {
+	/// Contributors.
+	pub contributors: Vec<RemoteContributor>,
+}
+
+/// Representation of a remote contributor.
+#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
+pub struct RemoteContributor {
+	/// Username.
+	pub username:      Option<String>,
+	/// Title of the pull request.
+	pub pr_title:      Option<String>,
+	/// The pull request that the user created.
+	pub pr_number:     Option<i64>,
+	/// Labels of the pull request.
+	pub pr_labels:     Vec<String>,
+	/// Whether if the user contributed for the first time.
+	pub is_first_time: bool,
+}
+
+impl Hash for RemoteContributor {
+	fn hash<H: Hasher>(&self, state: &mut H) {
+		self.username.hash(state);
+	}
+}
+
+/// Creates a HTTP client for the remote.
+fn create_remote_client(
+	remote: &Remote,
+	accept_header: &str,
+) -> Result<ClientWithMiddleware> {
+	if !remote.is_set() {
+		return Err(Error::RemoteNotSetError);
+	}
+	let mut headers = HeaderMap::new();
+	headers.insert(
+		reqwest::header::ACCEPT,
+		HeaderValue::from_str(accept_header)?,
+	);
+	if let Some(token) = &remote.token {
+		headers.insert(
+			reqwest::header::AUTHORIZATION,
+			format!("Bearer {}", token.expose_secret()).parse()?,
+		);
+	}
+	headers.insert(reqwest::header::USER_AGENT, USER_AGENT.parse()?);
+	let client = Client::builder()
+		.timeout(Duration::from_secs(REQUEST_TIMEOUT))
+		.tcp_keepalive(Duration::from_secs(REQUEST_KEEP_ALIVE))
+		.default_headers(headers)
+		.build()?;
+	let client = ClientBuilder::new(client)
+		.with(Cache(HttpCache {
+			mode:    CacheMode::Default,
+			manager: CACacheManager {
+				path: dirs::cache_dir()
+					.ok_or_else(|| {
+						Error::DirsError(String::from(
+							"failed to find the user's cache directory",
+						))
+					})?
+					.join(env!("CARGO_PKG_NAME")),
+			},
+			options: HttpCacheOptions::default(),
+		}))
+		.build();
+	Ok(client)
+}
+
+pub(crate) trait RemoteClient {
+	/// Returns the API url.
+	fn api_url() -> String;
+
+	/// Returns the remote repository information.
+	fn remote(&self) -> Remote;
+
+	/// Returns the HTTP client for making requests.
+	fn client(&self) -> ClientWithMiddleware;
+
+	/// Retrieves a single page of entries.
+	async fn get_entries_with_page<T: DeserializeOwned + RemoteEntry>(
+		&self,
+		project_id: i64,
+		page: i32,
+	) -> Result<Vec<T>> {
+		let url = T::url(project_id, &Self::api_url(), &self.remote(), page);
+		debug!("Sending request to: {url}");
+		let response = self.client().get(&url).send().await?;
+		let response_text = if response.status().is_success() {
+			let text = response.text().await?;
+			trace!("Response: {:?}", text);
+			text
+		} else {
+			let text = response.text().await?;
+			error!("Request error: {}", text);
+			text
+		};
+		let response = serde_json::from_str::<Vec<T>>(&response_text)?;
+		if response.is_empty() {
+			Err(Error::PaginationError(String::from("end of entries")))
+		} else {
+			Ok(response)
+		}
+	}
+
+	/// Fetches the remote API and returns the given entry.
+	async fn fetch<T: DeserializeOwned + RemoteEntry>(
+		&self,
+		project_id: i64,
+	) -> Result<Vec<T>> {
+		let entries: Vec<Vec<T>> = stream::iter(0..)
+			.map(|i| self.get_entries_with_page(project_id, i))
+			.buffered(T::buffer_size())
+			.take_while(|page| {
+				if let Err(e) = page {
+					debug!("Error while fetching page: {:?}", e);
+				}
+				future::ready(page.is_ok())
+			})
+			.map(|page| match page {
+				Ok(v) => v,
+				Err(ref e) => {
+					log::error!("{:#?}", e);
+					page.expect("failed to fetch page: {}")
+				}
+			})
+			.collect()
+			.await;
+		Ok(entries.into_iter().flatten().collect())
+	}
+
+	/// Retrieves a single object.
+	async fn get_entry<T: DeserializeOwned + RemoteEntry>(&self) -> Result<T> {
+		let url = T::url(0, &Self::api_url(), &self.remote(), 1);
+		debug!("Sending request to: {url}");
+		let response = self.client().get(&url).send().await?;
+		let response_text = if response.status().is_success() {
+			let text = response.text().await?;
+			trace!("Response: {:?}", text);
+			text
+		} else {
+			let text = response.text().await?;
+			error!("Request error: {}", text);
+			text
+		};
+		Ok(serde_json::from_str::<T>(&response_text)?)
+	}
+}
 
 /// Generates a function for updating the release metadata for a remote.
 #[macro_export]
